+++ src/scripts/settings.ts
@@ -5,12 +5,12 @@ import weather from './features/weather'
 import searchbar from './features/searchbar'
 import customFont from './features/fonts'
 import quickLinks from './features/links'
-import linksImport, { syncNewBookmarks } from './features/linksImport'
 import hideElements from './features/hide'
 import moveElements from './features/move'
 import localBackgrounds from './features/localbackgrounds'
 import unsplashBackgrounds from './features/unsplash'
 import storage, { getSyncDefaults } from './storage'
+import linksImport, { syncNewBookmarks } from './features/links/bookmarks'
 
 import langList from './langs'
 import parse from './utils/parse'
